moved create transaction dialog to component
